py: libcamera: Improve python binding installation
The existing meson.build file installs the bindings to an architecture
specific libdir (eg /usr/local/lib/aarch64-linux-gnu/), which is not
picked up by default python which only looks in the non architecture
specific libdir (eg /usr/local/lib/python3.11/). It also will always
build using the system python, rather than building using the same
python as meson is using. This prevents a user being able to build the
bindings for a different version of python, without changing their
system python to that version.

Modify the build process to use the meson Python module to build the
python bindings targets, so it installs them to the correct directories
for python, and builds them for the version of python that meson is
running with. For cross-compiling, still use the previous method to
build the bindings, as the host machine version of python should be
used instead.

Signed-off-by: William Vinnicombe <william.vinnicombe@raspberrypi.com>